pingpong (easy)
Write a program that uses UNIX system calls to ''ping-pong'' a byte
between two processes over a pair of pipes, one for each direction. The
parent should send a byte to the child; the child should print "<pid>:
received ping", where <pid> is its process ID, write the byte on the
pipe to the parent, and exit; the parent should read the byte from the
child, print "<pid>: received pong", and exit. Your solution should be
in the file user/pingpong.c

Some hints:

Use pipe to create a pipe.
Use fork to create a child.
Use read to read from a pipe, and write to write to a pipe.
Use getpid to find the process ID of the calling process.
Add the program to UPROGS in Makefile.

User programs on xv6 have a limited set of library functions available to
them. You can see the list in user/user.h; the source (other than for
system calls) is in user/ulib.c, user/printf.c, and user/umalloc.c.

Run the program from the xv6 shell and it should produce the following output:

    $ make qemu
    ...
    init: starting sh
    $ pingpong
    4: received ping
    3: received pong
    $

Your solution is correct if your program exchanges a byte between two
processes and produces output as shown above.

Tags: