Sysinfo (moderate)
In this assignment you will add a system call, sysinfo, that collects
information about the running system. The system call takes one argument:
a pointer to a struct sysinfo (see kernel/sysinfo.h). The kernel should
fill out the fields of this struct: the freemem field should be set to
the number of bytes of free memory, and the nproc field should be set to
the number of processes whose state is not UNUSED. We provide a test
program sysinfotest; you pass this assignment if it prints "sysinfotest: OK".

Some hints:

Add $U/_sysinfotest to UPROGS in Makefile

Run make qemu; user/sysinfotest.c will fail to compile. Add the system
call sysinfo, following the same steps as in the previous assignment.
To declare the prototype for sysinfo() in user/user.h you need
predeclare the existence of struct sysinfo:

    struct sysinfo;
    int sysinfo(struct sysinfo *);

Once you fix the compilation issues, run sysinfotest; it will fail
because you haven't implemented the system call in the kernel yet.
sysinfo needs to copy a struct sysinfo back to user space;
see sys_fstat() (kernel/sysfile.c) and filestat() (kernel/file.c)
for examples of how to do that using copyout().

To collect the amount of free memory, add a function to kernel/kalloc.c

To collect the number of processes, add a function to kernel/proc.c