Detect which pages have been accessed (hard)
Some garbage collectors (a form of automatic memory management)
can benefit from information about which pages have been accessed
(read or write). In this part of the lab, you will add a new
feature to xv6 that detects and reports this information to
userspace by inspecting the access bits in the RISC-V page table.
The RISC-V hardware page walker marks these bits in the PTE
whenever it resolves a TLB miss.

Your job is to implement pgaccess(), a system call that reports
which pages have been accessed. The system call takes three
arguments. First, it takes the starting virtual address of the
first user page to check. Second, it takes the number of pages
to check. Finally, it takes a user address to a buffer to store
the results into a bitmask (a datastructure that uses one bit per
page and where the first page corresponds to the least significant
bit). You will receive full credit for this part of the lab if the
pgaccess test case passes when running pgtbltest.
Some hints:

Read pgaccess_test() in user/pgtlbtest.c to see how pgaccess is used.
Start by implementing sys_pgaccess() in kernel/sysproc.c.
You'll need to parse arguments using argaddr() and argint().
For the output bitmask, it's easier to store a temporary buffer in
the kernel and copy it to the user (via copyout()) after filling it
with the right bits.

It's okay to set an upper limit on the number of pages that can be
scanned.
walk() in kernel/vm.c is very useful for finding the right PTEs.
You'll need to define PTE_A, the access bit, in kernel/riscv.h.
Consult the RISC-V privileged architecture manual to determine its value.
Be sure to clear PTE_A after checking if it is set. Otherwise, it won't
be possible to determine if the page was accessed since the last time
pgaccess() was called (i.e., the bit will be set forever).
vmprint() may come in handy to debug page tables.